restructuring medtronic folders
