+++ lib/drivers/medtronic/cli/blob_loader.js
@@ -3,13 +3,13 @@
 var program = require('commander');
 var fs = require('fs');
 
-var driverManager = require('../../driverManager');
+var driverManager = require('../../../driverManager');
 var pwdSimulator = require('../medtronicSimulator.js');
-var builder = require('../../objectBuilder.js')();
-var medtronicDriver = require('../../drivers/medtronicDriver');
-var api = require('../../core/api.js');
-var config = require('../../../.config.js');
-var pkg = require('./../../../package.json');
+var builder = require('../../../objectBuilder.js')();
+var medtronicDriver = require('../medtronicDriver');
+var api = require('../../../core/api.js');
+var config = require('../../../../.config.js');
+var pkg = require('./../../../../package.json');
 
 var intro = 'Medtronic CLI:';
