+++ lib/drivers/medtronic/docs/images/src/basal.diag
