+++ lib/drivers/medtronic/docs/images/src/basalSchedules.diag
