+++ lib/drivers/medtronic/docs/images/src/lowBattery.diag
