+++ lib/drivers/medtronic/docs/images/src/lowReservoir.diag
