+++ lib/drivers/medtronic/docs/images/src/prime.diag
