+++ lib/drivers/medtronic/docs/images/src/resume.diag
