+++ lib/drivers/medtronic/docs/images/src/smbg.diag
