+++ lib/drivers/medtronic/docs/images/src/smbgLinked.diag
