+++ lib/drivers/medtronic/docs/images/src/suspend.diag
