+++ lib/drivers/medtronic/docs/images/src/wizard.diag
