+++ lib/drivers/medtronic/docs/images/src/wizardSettings.diag
