+++ lib/drivers/medtronic/medtronicDriver.js
@@ -20,15 +20,15 @@
 var _ = require('lodash');
 var async = require('async');
 var sundial = require('sundial');
-var struct = require('../struct.js')();
-var annotate = require('../eventAnnotations');
-var proc = require('../medtronic/processData');
-var common = require('../commonFunctions');
-var medtronicSimulator = require('../medtronic/medtronicSimulator');
-var crcCalculator = require('../crc.js');
+var struct = require('../../struct.js')();
+var annotate = require('../../eventAnnotations');
+var proc = require('./processData');
+var common = require('../../commonFunctions');
+var medtronicSimulator = require('./medtronicSimulator');
+var crcCalculator = require('../../crc.js');
 
 var isBrowser = typeof window !== 'undefined';
-var debug = isBrowser ? require('../bows')('MedtronicDriver') : console.log;
+var debug = isBrowser ? require('../../bows')('MedtronicDriver') : console.log;
 
 module.exports = function (config) {
   var cfg = _.clone(config);
