+++ lib/drivers/medtronic/medtronicSimulator.js
@@ -19,10 +19,10 @@ var _ = require('lodash');
 var sundial = require('sundial');
 var util = require('util');
 
-var annotate = require('../eventAnnotations');
+var annotate = require('../../eventAnnotations');
 
-var debug = require('../bows')('MedtronicDriver');
-var common = require('../commonFunctions');
+var debug = require('../../bows')('MedtronicDriver');
+var common = require('../../commonFunctions');
 
 /**
  * Creates a new "simulator" for Medtronic pump data.
