+++ lib/drivers/medtronic/processData.js
@@ -20,13 +20,13 @@
 var _ = require('lodash');
 
 var sundial = require('sundial');
-var struct = require('../struct.js')();
-var common = require('../commonFunctions');
-var annotate = require('../eventAnnotations');
-var TZOUtil = require('../TimezoneOffsetUtil');
+var struct = require('../../struct.js')();
+var common = require('../../commonFunctions');
+var annotate = require('../../eventAnnotations');
+var TZOUtil = require('../../TimezoneOffsetUtil');
 
 var isBrowser = typeof window !== 'undefined';
-var debug = isBrowser ? require('../bows')('MedtronicDriver') : console.log;
+var debug = isBrowser ? require('../../bows')('MedtronicDriver') : console.log;
 
 var RECORD_TYPES = {
   BOLUS: { value: 0x01, head_length: 8, name: 'BOLUS'},
